@@ -4,16 +4,20 @@
   "version": "1.0.0",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "dependencies": {
-    "gatsby": "^1.9.277",
-    "gatsby-link": "^1.6.40",
-    "gatsby-plugin-react-helmet": "^2.0.10",
-    "gatsby-plugin-react-next": "^1.0.11",
-    "gatsby-plugin-sass": "^1.0.26",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-helmet": "^5.2.0",
-    "react-spring": "^5.4.0",
-    "simple-modular-scale": "^1.0.2"
+    "framer-motion": "^10.16.3",
+    "gatsby": "^5.12.3",
+    "gatsby-link": "^5.12.0",
+    "gatsby-plugin-image": "^3.12.0",
+    "gatsby-plugin-postcss": "^6.12.0",
+    "gatsby-plugin-sass": "^6.12.0",
+    "gatsby-plugin-sharp": "^5.12.0",
+    "gatsby-source-filesystem": "^5.12.0",
+    "gatsby-transformer-sharp": "^5.12.0",
+    "modular-scale": "^5.1.2",
+    "postcss": "^8.4.29",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-spring": "^9.7.2"
   },
   "keywords": [
     "gatsby"
@@ -26,9 +30,8 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "devDependencies": {
-    "bootstrap": "^4.0.0-beta",
-    "modularscale-sass": "^3.0.5",
-    "prettier": "^1.12.0"
+    "bootstrap": "^5.3.1",
+    "prettier": "^3.0.3"
   },
   "repository": {
     "type": "git",
