@@ -3,13 +3,28 @@ module.exports = {
     title: 'Lighthouse & Giant',
   },
   plugins: [
-    {
-      resolve: `gatsby-plugin-react-helmet`,
-    },
+    `gatsby-plugin-image`,
+    `gatsby-plugin-sharp`,
+    `gatsby-transformer-sharp`, // Needed for dynamic images
     {
       resolve: `gatsby-plugin-sass`,
       options: {
-        precision: 8,
+        cssLoaderOptions: {
+          esModule: false,
+          modules: {
+            namedExport: false,
+          },
+        },
+      },
+    },
+    `gatsby-plugin-postcss`,
+    {
+      resolve: `gatsby-source-filesystem`,
+      options: {
+        // The unique name for each instance
+        name: `images`,
+        // Path to the directory
+        path: `${__dirname}/src/images/`,
       },
     },
   ],
