@@ -1,7 +1,24 @@
-/**
- * Implement Gatsby's SSR (Server Side Rendering) APIs in this file.
- *
- * See: https://www.gatsbyjs.org/docs/ssr-apis/
- */
+// gatsby-ssr.js
 
- // You can delete this file if you're not using it
+import React from 'react'
+
+export const onRenderBody = ({ setHeadComponents }) => {
+  setHeadComponents([
+    <link
+      key="humane-semiBold"
+      rel="preload"
+      href="../fonts/Humane-SemiBold.woff2"
+      as="font"
+      type="font/woff2"
+      crossOrigin="anonymous"
+    />,
+    <link
+      key="humane-regular"
+      rel="preload"
+      href="../fonts/Humane-Regular.woff2"
+      as="font"
+      type="font/woff2"
+      crossOrigin="anonymous"
+    />,
+  ])
+}
