@@ -1,51 +1,96 @@
 import React from 'react'
-import Link from 'gatsby-link'
-import './font.css'
-import './aboutTile.scss'
+import { Link } from 'gatsby'
+import { motion } from 'framer-motion'
 
-const ms = require('simple-modular-scale')
+const AboutTile = ({ siteTitle }) => {
+  const container = {
+    hidden: { opacity: 0 },
+    show: {
+      opacity: 1,
+      transition: {
+        delay: 0.2,
+        staggerChildren: 0.3,
+      },
+    },
+  }
 
-const scale = ms({
-  base: 16,
-  ratios: [3 / 2, 4 / 3],
-  length: 8,
-})
-// [ 16, 24, 32, 48, 64, 96, 128, 192 ]
-
-const AboutTile = ({ siteTitle }) => (
-  <section
-    className="about"
-    style={{
-      marginBottom: scale[1],
-      paddingTop: scale[0],
-      paddingBottom: '1.45rem',
-      position: 'fixed',
-      zIndex: '9999999999',
-      display: 'flex',
-      alignItems: 'center',
-      justifyContent: 'center',
-      width: '100vw',
-      height: '100vh',
-    }}
-  >
-    <h1>Lighthouse &amp; Giant</h1>
-    <p>
-      Back in 2010 we started a psuedo agency, something to pour our creativity
-      into to try and help build brands. We've worked with mostly startups but
-      we like to think we've helped them on their way to become the mature
-      brands they are today, head held high. Fast forward to 2017 and we've
-      joined an real life agency.<br />This is where you can now find us.
-    </p>
-    <Link
+  const item = {
+    hidden: { opacity: 0, y: '1rem' },
+    show: { opacity: 1, y: 0 },
+  }
+  return (
+    <section
+      className="about"
       style={{
-        textDecoration: 'none',
-        fontSize: scale[2],
+        paddingBottom: '1.45rem',
+        position: 'fixed',
+        zIndex: '9999999999',
+        display: 'flex',
+        flexDirection: 'column',
+        alignItems: 'center',
+        justifyContent: 'center',
+        width: '100vw',
+        height: '100vh',
+        paddingLeft: '2rem',
+        paddingRight: '2rem',
       }}
-      to="/"
     >
-      &times;
-    </Link>
-  </section>
-)
+      <motion.div
+        variants={container}
+        initial="hidden"
+        animate="show"
+        style={{
+          maxWidth: 600,
+          marginLeft: 'auto',
+          marginRight: 'auto',
+          textAlign: 'center',
+        }}
+      >
+        <motion.div variants={item}>
+          <h1
+            style={{
+              fontFamily: 'Humane-SemiBold',
+              fontSize: '4rem',
+              textTransform: 'uppercase',
+              letterSpacing: '1px',
+            }}
+          >
+            Lighthouse &amp; Giant
+          </h1>
+        </motion.div>
+        <motion.div variants={item}>
+          <p
+            style={{
+              marginBottom: '1rem',
+              fontFamily: 'Nohemi-Regular',
+              fontSize: '1rem',
+              lineHeight: 1.8,
+            }}
+          >
+            Back in 2010 we started a psuedo agency, something to pour our
+            creativity into to try and help build brands. We've worked with
+            mostly startups but we like to think we've helped them on their way
+            to become the mature brands they are today, head held high. Fast
+            forward to 2017 and we've joined an real life agency.
+            <br />
+            <br />
+            This is where you can now find us.
+          </p>
+        </motion.div>
+        <motion.div variants={item}>
+          <Link
+            style={{
+              textDecoration: 'none',
+              // fontSize: ms(2),
+            }}
+            to="/"
+          >
+            &times;
+          </Link>
+        </motion.div>
+      </motion.div>
+    </section>
+  )
+}
 
 export default AboutTile
